rag-backend/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.34.0
cohere==5.11.2
qdrant-client==1.12.1
pydantic==2.10.3
python-dotenv==1.0.1
pyyaml==6.0.2
playwright==1.48.0
beautifulsoup4==4.12.3
trafilatura==1.13.0